+++ src/main/java/com/kb/csrm/course/service/impl/CourseServiceImpl.java
@@ -30,7 +30,7 @@ public class CourseServiceImpl implements ICourseService {
     }
 
     @Override
-    public CourseDto selectCourseById(int courseId) {
+    public CourseDto selectCourseById(Long courseId) {
         return courseMapper.selectCourseById(courseId);
     }
